+++ firehosepool/connect.go
@@ -2,7 +2,6 @@ package firehosepool
 
 import (
 	"fmt"
-	"log"
 	"time"
 )
 
@@ -19,16 +18,11 @@ func (srv *Server) _reload() {
 		if srv.isExiting() {
 			continue
 		}
-
-		if err := srv.clientsReset(); err != nil {
-			log.Printf("Firehose ERROR: can't connect to kinesis: %s", err)
-			time.Sleep(connectionRetry)
-		}
+		srv.clientsReset()
 	}
 }
 
 func (srv *Server) failure() {
-	// log.Println("MARK FAILURE ...")
 	if srv.isExiting() {
 		return
 	}
@@ -42,7 +36,7 @@ func (srv *Server) failure() {
 	srv.lastError = time.Now()
 	srv.Unlock()
 
-	log.Printf("Firehose: %d errors detected", srv.errors)
+	logger.Info("Streamspooler failure marked", "stream", srv.cfg.StreamName, "errors", srv.errors, "reload", reload)
 
 	if reload {
 		select {
@@ -54,7 +48,7 @@ func (srv *Server) failure() {
 
 func (srv *Server) fhClientReset(cfg *Config) (err error) {
 	if srv.lastConnection.Add(limitIntervalConnection).Before(time.Now()) {
-		log.Printf("creating new firehose client for the stream %s ...", srv.cfg.StreamName)
+		logger.Info("Streamspooler: creating new aws firehose client", "stream", srv.cfg.StreamName)
 
 		var fhcg ClientGetter
 		if cfg.FHClientGetter != nil {
@@ -75,20 +69,19 @@ func (srv *Server) fhClientReset(cfg *Config) (err error) {
 	return
 }
 
-func (srv *Server) clientsReset() (err error) {
+func (srv *Server) clientsReset() {
 	srv.Lock()
 	defer srv.Unlock()
-	log.Println("Firehose: RELOADING clients")
 
 	defer func() {
-		log.Printf("Firehose %s clients %d, in the queue %d/%d", srv.cfg.StreamName, len(srv.clients), len(srv.C), cap(srv.C))
+		logger.Info("Streamspooler: workers reset done", "stream", srv.cfg.StreamName, "workers", len(srv.clients), "in-queue", fmt.Sprintf("%d/%d", len(srv.C), cap(srv.C)))
 	}()
 
 	currClients := len(srv.clients)
 
 	// No changes in the number of clients
 	if currClients == srv.cliDesired {
-		return nil
+		return
 	}
 
 	// If the config define lower number than the active clients remove the difference
@@ -105,6 +98,4 @@ func (srv *Server) clientsReset() (err error) {
 			srv.clients = append(srv.clients, NewClient(srv))
 		}
 	}
-
-	return nil
 }